backend/requirements.txt
google-generativeai 
langchain-google-genai
langchain

scikit-learn
torch
numpy

fastapi
transformers
chromadb==0.4.3

pypdf
docx2txt
unstructured
unstructured[pdf]
tqdm